chore: upgrade rudder-iac to v0.22.0
Move the provider off the pinned pre-release rudder-iac commit (#625) to the
released v0.22.0, which is required to access the destination-side transformation
endpoints used by the destination↔transformation link feature.

The upgrade carries a few mechanical, behaviour-preserving adaptations to
rudder-iac API changes between the old pin and v0.22.0:

- retl: CreateRETLConnectionRequest.SyncBehaviour became *SyncBehaviour — take
  the address (value, including "", is still sent). Test literals wrapped via a
  ptr() helper.
- generatetf: Destination.Version widened to int64 — cast at the call sites.
- destination: populate Destination.Version via int64(cm.Version); existing
  version tests assert int64.
- testutil/acc: compare against int64(wantVersion).

No behavioural change; build + unit tests green.


🔒 Scanned for secrets using gitleaks 8.28.0

+++ internal/testutil/acc/destinations.go
@@ -201,7 +201,7 @@ func testAccCheckDestinationVersion(resourceName string, wantVersion int) resour
 			return fmt.Errorf("failed to get destination from API: %w", err)
 		}
 
-		if dest.Version != wantVersion {
+		if dest.Version != int64(wantVersion) {
 			return fmt.Errorf("destination %s: got version %d, want %d", rs.Primary.ID, dest.Version, wantVersion)
 		}
 		return nil
